+++ lib/galaxy/model/database_heartbeat.py
