+++ scripts/galaxy-main
@@ -54,7 +54,7 @@ except Exception:
     log.exception("Failed to add Galaxy to sys.path")
     raise
 from galaxy.util import unicodify
-from galaxy.web.stack import get_app_kwds
+from galaxy.web_stack import get_app_kwds
 
 REQUIRES_DAEMONIZE_MESSAGE = "Attempted to use Galaxy in daemon mode, but daemonize is unavailable."
